refactor: split gui.h into lightweight public header and implementation header
