+++ src/gui/content_pages/page_optical_system_info.cpp
@@ -1,4 +1,5 @@
 #include "gui/gui.h"
+#include "lib/imgui/imgui.h"
 
 namespace gui {
     void GuiManager::renderPageOpticalSystemInfo() {
